requirements.txt
# TechFilings 依赖

# 网络请求
requests>=2.28.0

# 文档解析
beautifulsoup4>=4.11.0
lxml>=4.9.0

# 向量数据库
chromadb>=0.4.0

# LLM和Embedding
openai>=1.0.0
llama-index>=0.9.0

# Web界面
streamlit>=1.28.0

# 工具
python-dotenv>=1.0.0
tqdm>=4.65.0

tiktoken>=0.1.0